AJUSTES REVISION 2

+++ src/app/layout/header/header.component.ts
@@ -3,6 +3,7 @@ import { Router } from '@angular/router';
 import { faBars, faBell, faComment } from '@fortawesome/free-solid-svg-icons';
 import { VacationService } from '@pages/vacation/vacation.service';
 import { CookieService } from 'ngx-cookie-service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-header',
@@ -23,19 +24,32 @@ public logo = 'assets/images/osinermin.png';
   ngOnInit(): void {
   }
 
-  async logout() {
-    await new Promise((resolve, reject) => {
-      try {
-        this.cookieService.delete('isLoggedIn', '/')
-        this.cookieService.delete('identificacion', '/')
-        resolve(true)
+  logout() {
+    Swal.fire({
+      title: 'Esta seguro de cerrar la sesion?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Cerrar sesion',
+      cancelButtonText: 'Cancelar',
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        await new Promise((resolve, reject) => {
+          try {
+            this.cookieService.delete('isLoggedIn', '/')
+            this.cookieService.delete('identificacion', '/')
+            resolve(true)
+          }
+          catch (err) {
+            reject(false)
+          }
+    
+        }).then(() => {
+          this.router.navigateByUrl('/home')
+        })
+       
       }
-      catch (err) {
-        reject(false)
-      }
-
-    }).then(() => {
-      this.router.navigateByUrl('/home')
     })
   }
 }
